### 2025-04-02 #7987 @delcmo
List of reviewers:
- [Sven Berger](https://reviewers.joss.theoj.org/reviewers/2208)
- [Khemraj Gautam Kshetri](https://reviewers.joss.theoj.org/reviewers/13671)
- [Martin E. Liza](https://reviewers.joss.theoj.org/reviewers/7995)
- [pratikvn](https://reviewers.joss.theoj.org/reviewers/1112)
- [Stefan Zitz](https://reviewers.joss.theoj.org/reviewers/2061)

Let me know if you need additional names.

@arfon I will merge `main` (default) to `gh-pages`. I was not aware the source code the documentation branch had to be in a single branch.
