### 2025-07-01 #7987 @delcmo
I think I was asked to merge master into the gh-pages branch by a JOSS
person. It is fine to review the source code but it will take time.

Marc-Olivier Delchini

On Tue, Jul 1, 2025 at 2:00 AM Pi-Yueh Chuang ***@***.***>
wrote:

> *piyueh* left a comment (openjournals/joss-reviews#7987)
> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3021937240>
>
> Hi @delcmo <https://github.com/delcmo>, I’ll be editing this submission.
> I noticed you merged the master branch into the gh-pages branch, but your
> submission indicates you want the gh-papers branch to be reviewed. Did
> you mean to merge master into gh-papers instead? Please note that the
> reviewers will review the source code, so the branch under review must
> include both the source code and the JOSS paper.
>
