### 2025-07-01 #7987 @piyueh
Thanks for clarifying, @kyleniemeyer!

@delcmo, I think there may have been some confusion in your previous conversation between `gh-papers` and `gh-pages` (since your repository has both branches). But in any case, since it’s been clarified that `paper.md` and the source code don’t need to be in the same branch, I’ll proceed and assume the code to review is in `master`. Please let me know if that’s not correct.
