### 2025-07-01 #7987 @delcmo
This is correct. The code in master can be reviewed. Thanks for checking
with me.

On Tue, Jul 1, 2025 at 12:26 PM Pi-Yueh Chuang ***@***.***>
wrote:

> *piyueh* left a comment (openjournals/joss-reviews#7987)
> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3024729848>
>
> Thanks for clarifying, @kyleniemeyer <https://github.com/kyleniemeyer>!
>
> @delcmo <https://github.com/delcmo>, I think there may have been some
> confusion in your previous conversation between gh-papers and gh-pages
> (since your repository has both branches). But in any case, since it’s been
> clarified that paper.md and the source code don’t need to be in the same
> branch, I’ll proceed and assume the code to review is in master. Please
> let me know if that’s not correct.
>
