### 2025-07-01 #7987 @xuanxu
It looks like the repository is too big, [the error](https://github.com/openjournals/joss-papers/actions/runs/16006834059/job/45155362654#step:2:116) mentions `This repository exceeded its LFS budget`
