### 2025-07-01 #7987 @delcmo
Is that master or gh-papers branch?

Marc-Olivier Delchini

On Tue, Jul 1, 2025 at 2:09 PM Juanjo Bazán ***@***.***>
wrote:

> *xuanxu* left a comment (openjournals/joss-reviews#7987)
> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3025042404>
>
> It looks like the repository is too big, the error
> <https://github.com/openjournals/joss-papers/actions/runs/16006834059/job/45155362654#step:2:116>
> mentions This repository exceeded its LFS budget
>
