### 2025-07-01 #7987 @delcmo
I see the same issues but not sure what is causing it. It was working fine
last time I checked a couple of weeks ago. I was able to clone all files
but LFS files with `GIT_LFS_SKIP_SMUDGE=1 git clone
https://github.com/ORNL/VERTEX-CFD`. LFS files are Exodus files (solution
file or mesh file).

Marco

On Tue, Jul 1, 2025 at 3:10 PM Pi-Yueh Chuang ***@***.***>
wrote:

> *piyueh* left a comment (openjournals/joss-reviews#7987)
> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3025213866>
>
> Thanks for looking into it @xuanxu <https://github.com/xuanxu> .
>
> @delcmo <https://github.com/delcmo> It's master. I tested it on my local
> machine and also got the same LFS budget error. Steps to reproduce:
>
>    1. git clone https://github.com/ORNL/VERTEX-CFD
>    2. cd VERTEX-CFD
>    3. git lfs pull
>
