### 2025-07-03 #7987 @piyueh
Hi @delcmo, are these mesh files required for running tests and examples? Reviewers will need to run tests and examples in order to review your software. So, we may need to wait until you resolve the LFS issues if these mesh files are required.
