### 2025-07-03 #7987 @delcmo
Not all input files require an Exodus mesh file. Some of them rely on
INLINE mesh capabilities, which is a Panzer functionality. You should be
able to compile the code and run any input files with the INLINE mesh
block. You will not be able to run regression tests though since all gold
files (Exodus file) are LFS files. I will work on this today or tomorrow.
It seems to be a GiHub issue.

Marco

On Thu, Jul 3, 2025 at 10:01 AM Pi-Yueh Chuang ***@***.***>
wrote:

> *piyueh* left a comment (openjournals/joss-reviews#7987)
> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3032393807>
>
> Hi @delcmo <https://github.com/delcmo>, are these mesh files required for
> running tests and examples? Reviewers will need to run tests and examples
> in order to review your software. So, we may need to wait until you resolve
> the LFS issues if these mesh files are required.
>
