### 2025-07-16 39ebb27
removing exodus files
