### 2025-07-16 #7987 @delcmo
I was able to think of a workaround. I removed all LFS files from the
GitHub repository and moved them to a public GitLab repository:
https://code.ornl.gov/mxd/vertex-cfd.git

You should be able to git clone the VERTEX-CFD repo from GitHub with `git
clone https://github.com/ORNL/VERTEX-CFD.git .` and get the input files,
mesh files, and gold files for the regression test from the public GItLab
repo: `git clone https://code.ornl.gov/mxd/vertex-cfd.git .`. You may have
to switch to the `master` branch.

Note that the `examples` and `regression_test` folders will have to be
moved to the GitHub repo locally to recover the original set up (same level
as the `src` directory).

Let me know if you have any questions.

Marco

On Sat, Jul 12, 2025 at 9:55 AM Marco Delchini ***@***.***>
wrote:

> I am trying to resolve the issue but it will most likely require me to
> remove all LFS files.
>
> Marc-Olivier Delchini
>
>
> On Fri, Jul 11, 2025 at 11:26 AM Pi-Yueh Chuang ***@***.***>
> wrote:
>
>> *piyueh* left a comment (openjournals/joss-reviews#7987)
>> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3062737154>
>>
>> Hi @delcmo <https://github.com/delcmo> Just checking if the LFS problem
>> has been resolved? The LFS problem is blocking this submission and the
>> reviewing process from moving forward.
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-3062737154>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ABD4GCSHP6K4QT6SYDWWZ2T3H7JSDAVCNFSM6AAAAAB2JKQE42VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTANRSG4ZTOMJVGQ>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>

--
Marc-Olivier Delchini
