### 2025-07-16 #7987 @piyueh
Thanks for the update @delcmo . Separating the data files (mesh, etc.) from the source code is fine for JOSS, as long as your documentation has enough step-by-step instructions for reviewers to download these data files and to run tests and/or benchmark cases. So I will now re-run editorial checks, and if the editorial checks pass, I will start the process of finding reviewers.
