### 2025-07-21 #7987 @piyueh
Hi @Zitzeronion @ma-sadeghi @anand-me @smillerc @sthavishtha @Phoenixfire1081 @andrewgiuliani @xcheng86 @jamesafful @khemgautam27 @knstmrd Thank you for signing up as JOSS reviewers. Would you be willing to review this JOSS submission? A draft PDF of this submission can be found [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07987/joss.07987/10.21105.joss.07987.pdf). The source code can be found [here](https://github.com/ORNL/VERTEX-CFD). The examples and tests can be found [here](https://code.ornl.gov/mxd/vertex-cfd/-/tree/master?ref_type=heads). If you are not familiar with JOSS review standard and process, we provide detailed guidelines to guide you how to review for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
