### 2025-07-22 d04ea42
Minor layouting thing
Added math layout to u_i^avg in the laminar flow section.
