### 2025-07-22 #7987 @Zitzeronion
@piyueh I think I am unqualified to review this contribution, I have very limited knowledge about MHD and or fission/fusion dynamics. That said, it surely looks like an interesting piece of software.
