### 2025-07-22 #7987 @piyueh
@Phoenixfire1081 @sthavishtha Thanks for agreeing to review. I want to wait for another two or three days and see if I can find one more reviewer. Two reviewers are the minimum requirement for JOSS, but we usually try to have three reviewers if possible. The review process will start once we finish the process of finding reviewers.

@Zitzeronion Thanks for your response. The author specifically recommended you (see [this comment](https://github.com/openjournals/joss-reviews/issues/7987#issuecomment-2773177795)), suggesting they believe you have sufficient domain knowledge and value your perspective. Additionally, it's not uncommon for JOSS reviewers to have limited domain knowledge since JOSS focuses more on proper practices for scientific software development and less on what applications the software is for. Let me know if you'd reconsider reviewing this submission.
