### 2025-07-28 #8654 @editorialbot
[REVIEW]: VERTEX-CFD: A multiphysics platform for fusion applications
**Submitting author:** @delcmo (<a href="http://orcid.org/0000-0003-3263-254X">Marc Olivier Delchini</a>)
**Repository:** https://github.com/ORNL/VERTEX-CFD
**Branch with paper.md** (empty if default branch): gh-papers
**Version:** v1.0.0
