### 2025-08-30 ORNL/VERTEX-CFD#9 @sthavishtha
JOSS article review #1
Please find my first set of comments here:

Comments on the software paper
1. Please verify the list of co-authors mentioned in the manuscript. As per JOSS guidelines (https://joss.readthedocs.io/en/latest/submitting.html#authorship), financial and organizational contributions cannot be considered for co-authorship of JOSS articles. Firstly, the first 3 authors listed in the JOSS article have contributed to the code, while the rest of the authors have not. Secondly, the 4th contributed author to the JOSS code (robertsrb) has not been listed as a co-author of the JOSS article.
2. It would be good to showcase the capability of this software by showing some simulation results/plots. Please see the software articles published in JOSS, and add some simulation plots accordingly. Also ensure that the data for these simulation results/plots is made available to the reviewer. Please see the reviewer checklist for more details (https://github.com/openjournals/joss-reviews/issues/8654#issuecomment-3218499457). Also ensure that these results are reproducible - the reviewers will verify that later.
3. In lines 43-56, the authors mention the need for using RANS turbulence models to simulate MHD turbulent flows. But, I do not find any mention of turbulence models in the software documentation. Does the software support RANS turbulence models or is it planned to include them in the future?
4. Line 58: what is the need for complex geometries while modeling MHD flows? Please add details in the paper.
5. Lines 59-60: the phrase "integrates AI and ML tools" is very broad. what kind of AI & ML tools can be integrated into the software? Please add details in the paper.
6. Lines 61-62: what is "design analysis"? Please replace this word with a more appropriate word in the paper.
7. **Statement of need**: There are in fact a few CPU and GPU portable HPC packages using FEM/spectral element methods (that i can recall), that can simulate a much larger range of multiphysics problems when compared to VERTEX-CFD: MFEM and Nek5000. In that case, what is the specific need for VERTEX-CFD? Isn’t it possible to use the other multiphysics codes (like MFEM and Nek5000) for modeling fusion applications? Please add a related discussion in the paper.
8. **State of the field**: no comparison is made between VERTEX-CFD and its software counterparts in the paper. Please add this discussion in the paper.
9. Disclosure/potential conflicts of interest (https://joss.readthedocs.io/en/latest/submitting.html#policies) are not reported in the article.
10. Please fix the following references:
- https://doi.org/10.1098/rspa.1937.0036 for “Mechanism of the production of small eddies….“
- please remove https://doi.org/ accidentally written in line 173
- doi in linę 147 is invalid

Once these comments are addressed, I will update my checklist (https://github.com/openjournals/joss-reviews/issues/8654#issuecomment-3218499457). My other set of comments will follow in the coming weeks.
