### 2025-09-01 83e6e58
Update paper.md
COI
