+++ paper/paper.bib
@@ -67,7 +67,7 @@ url = {https://trilinos.github.io}
 # TGV
 @article{Taylor-green-vortex,
  ISSN = {00804630},
- URL = {http://www.jstor.org/stable/96892},
+ URL = {https://doi.org/10.1098/rspa.1937.0036},
  author = {G. I. Taylor and A. E. Green},
  journal = {Proceedings of the Royal Society of London. Series A, Mathematical and Physical Sciences},
  number = {895},
@@ -218,6 +218,7 @@ abstract = {On the pathway toward full simulations for a liquid metal (LM) blank
   year={1992},
   url={https://api.semanticscholar.org/CorpusID:140573363}
 }
+
 
 @article{10.1023/a:1022818327584,
   author = {Castro, F. and Palma, J. M. L. M. and Lopes, A. S.},
