### 2025-09-01 841de45
Update paper.md
Update language on AI&ML
