### 2025-09-01 9d0876e
Update paper.md
Added content on complex geometry and AI and ML.
