### 2025-09-01 e4c5a47
Update paper.md
Added discussion on MFEM, Nek5000 and Trilinos
