### 2025-09-01 5ae0eb4
Update paper.bib
Updated references
