+++ paper/paper.bib
@@ -79,6 +79,33 @@ url = {https://trilinos.github.io}
  year = {1937}
 }
 
+@article{mfem,
+  title = {{MFEM}: A Modular Finite Element Methods Library},
+  author = {R. Anderson and J. Andrej and A. Barker and J. Bramwell and J.-S. Camier and
+            J. Cerveny and V. Dobrev and Y. Dudouit and A. Fisher and Tz. Kolev and W. Pazner and
+            M. Stowell and V. Tomov and I. Akkerman and J. Dahm and D. Medina and S. Zampini},
+  journal = {Computers \& Mathematics with Applications},
+  doi = {10.1016/j.camwa.2020.06.009},
+  volume = {81},
+  pages = {42-74},
+  year = {2021}
+}
+
+@article{nek5000
+    author = {Deville, , MO , Author and Fischer, , PF , Author and Mund , , EH , Author and Gartling, , DK , Reviewer},
+    title = {High-Order Methods for Incompressible Fluid Flow},
+    journal = {Applied Mechanics Reviews},
+    volume = {56},
+    number = {3},
+    pages = {B43-B43},
+    year = {2003},
+    month = {05},
+    issn = {0003-6900},
+    doi = {10.1115/1.1566402},
+    url = {https://doi.org/10.1115/1.1566402},
+    eprint = {https://asmedigitalcollection.asme.org/appliedmechanicsreviews/article-pdf/56/3/B43/5439822/b34\_1.pdf},
+}
+
 # Validation
 # Natural convection
 @article{Kuehn_Goldstein_1976, title={An experimental and theoretical study of natural convection in the annulus between horizontal concentric cylinders}, volume={74}, DOI={10.1017/S0022112076002012}, number={4}, journal={Journal of Fluid Mechanics}, author={Kuehn, T. H. and Goldstein, R. J.}, year={1976}, pages={695–719}} <div></div>
@@ -244,3 +271,4 @@ issn = "0003-018X",
 number = "1",
 note = "2024 Annual Conference on Transactions of the American Nuclear Society, ANS 2024 ; Conference date: 16-06-2024 Through 19-06-2024",
 }
+
