### 2025-09-01 9d36317
Update paper.md
Added references for MFEM and Nek5000
