### 2025-09-01 8c9a031
Update paper.md
