### 2025-09-01 ce3fa23
Update paper.bib
