+++ paper/paper.bib
@@ -92,7 +92,7 @@ url = {https://trilinos.github.io}
 }
 
 @article{nek5000
-    author = {Deville, , MO , Author and Fischer, , PF , Author and Mund , , EH , Author and Gartling, , DK , Reviewer},
+    author = {MO. Deville and PF. Fischer and EH. Mund and DK. Gartling},
     title = {High-Order Methods for Incompressible Fluid Flow},
     journal = {Applied Mechanics Reviews},
     volume = {56},
@@ -271,4 +271,5 @@ issn = "0003-018X",
 number = "1",
 note = "2024 Annual Conference on Transactions of the American Nuclear Society, ANS 2024 ; Conference date: 16-06-2024 Through 19-06-2024",
 }
+
 
