### 2025-09-01 d67895f
Update paper.bib
