+++ paper/paper.bib
@@ -91,7 +91,7 @@ url = {https://trilinos.github.io}
   year = {2021}
 }
 
-@article{nek5000
+@article{nek5000,
     author = {MO. Deville and PF. Fischer and EH. Mund and DK. Gartling},
     title = {High-Order Methods for Incompressible Fluid Flow},
     journal = {Applied Mechanics Reviews},
@@ -271,5 +271,6 @@ issn = "0003-018X",
 number = "1",
 note = "2024 Annual Conference on Transactions of the American Nuclear Society, ANS 2024 ; Conference date: 16-06-2024 Through 19-06-2024",
 }
+
 
 
