### 2025-09-01 b41bf03
Update paper.md
Added state of the art
