+++ paper/paper.md
@@ -49,6 +49,9 @@ bibliography: paper.bib
 
 The demand for high-performance computational fluid dynamics and multiphysics software packages has grown in recent years as a response to effort in complex engineering and research applications. Although the widespread deployment of high-performance computing (HPC) resources has enabled larger, more complex simulations to be conducted, few commercial or open-source software packages are available that (1) scale performantly on central processing unit (CPU) and graphical processing unit (GPU) computing architectures and (2) represent the multitude of physical processes relevant to these applications. The VERTEX-CFD software is developed to address this technical gap, with a special emphasis on high-fidelity multiphysics modeling of coupled turbulent fluid flow, heat transfer, and magnetohydrodynamics (MHD) for applications in fusion and fission energy and in other spaces. The VERTEX-CFD package was developed to solve the governing equations of these problems using a high-order continuous Galerkin finite element framework and fully implicit monolithic solvers. Special attention is being paid during the development process to verify and to validate the solver and to ensure performance portability across both CPU and GPU computing platforms. A comprehensive verification and validation (V&V) suite and unit tests were designed to assess the accuracy and convergence behavior of the VERTEX-CFD software for problems taken from the published literature.
 
+# State of the field
+
+Magnetohydrodynamic (MHD) simulations can be performed using both HIMAG and COMSOL Multiphysics, though each platform addresses different computational objectives. HIMAG is a specialized solver optimized for high-fidelity modeling of liquid-metal flows in complex geometries under strong magnetic fields, such as those encountered in fusion and industrial applications. Its design provides enhanced computational efficiency and accuracy for these niche MHD problems but limits flexibility for coupling with additional physical phenomena. In contrast, COMSOL Multiphysics is a general-purpose multiphysics environment that enables the simultaneous solution of fluid dynamics, electromagnetics, heat transfer, and structural mechanics. This versatility allows the simulation of a broad spectrum of MHD applications, including electromagnetic stirring, liquid-metal batteries, and reactor-relevant flows. Nevertheless, COMSOL’s generalist approach may present challenges in achieving numerical convergence, computational efficiency, and stable solver performance in extreme MHD regimes found in large Hartmann flows. Overall, HIMAG is particularly suited for targeted, high-fidelity fusion-scale MHD simulations, whereas COMSOL provides a flexible framework for investigating MHD phenomena within coupled multiphysics contexts. Additionally, demonstration of scalibility of HIMAG and COMSOL Multiphysics on GPU HPC platforms is yet to be demonstrated.
 
 # Statement of need
 
@@ -58,14 +61,13 @@ Two approaches are commonly adopted to predict MS flows exposed to a magnetic fi
 
 VERTEX-CFD is a new open-source package designed to address the modeling of MHD flows in complex geometries. It provides a robust multi-physics solver that can generate high-fidelity simulations by scaling on HPC platforms. Artificial intelligence and machine learning (AI&ML) capabilities are provided by TensorFlow Lite package [@tensorflow2015-whitepaper] that is coupled to VERTEX-CFD. The coupling allows to enhance physical models with ML models at runtime. These capabilities will provide the needed resources to extend the applicability of turbulence models to MHD flows, and perform modeling of complex flows under the influence of an external magnetic field. This in turn will be leveraged for the modeling of realistic complex geometries found in fusion blankets to capture the coupled multiphysics behavior that governs blanket performance and lifetime.
 
[…]
