### 2025-09-01 a0d465b
Update paper.bib
Added reference for HIMAG and COMSOL
