+++ paper/paper.bib
@@ -272,5 +272,27 @@ number = "1",
 note = "2024 Annual Conference on Transactions of the American Nuclear Society, ANS 2024 ; Conference date: 16-06-2024 Through 19-06-2024",
 }
 
+@article{HIMAG,
+title = {Validation strategies of HIMAG in interfacial flow computation for fusion applications},
+journal = {Fusion Engineering and Design},
+volume = {81},
+number = {8},
+pages = {1535-1541},
+year = {2006},
+note = {Proceedings of the Seventh International Symposium on Fusion Nuclear Technology},
+issn = {0920-3796},
+doi = {https://doi.org/10.1016/j.fusengdes.2005.09.053},
+url = {https://www.sciencedirect.com/science/article/pii/S0920379605006952},
+author = {Ming-Jiu Ni and Ramakanth Munipalli and Neil B. Morley and Mohamed A. Abdou},
+keywords = {Computational methods, MHD, Free surface, Level set, Validation},
+abstract = {HIMAG, a 3-D incompressible MHD free surface code developed by HyPerComp Inc., in a joint research project with UCLA has been successfully applied for APEX [M.A. Abdou, A. Ying, N. Morley, et al., On the exploration of innovative concepts for fusion chamber technology, APEX Interim Report Overview, Fusion Eng. Des. 54 (2001) 181–247] and fusion relevant cases [R. Munipalli, V. Shankar, M.-J. Ni, N. Morley, et al., Development of a 3-D incompressible free surface MHD computational environment for arbitrary geometries: HIMAG DOE phase-II SBIR, period of performance: June 2001–June 2003, Final Report, June 2003; N. Morley, S. Smolentsev, N. Munipalli, M.-J. Ni, D. Gao, M. Abdou, Progress on the modeling of liquid metal, free surface, MHD flows for fusion liquid walls, Fusion Eng. Des. 72 (2004) 3–34] including LIMITS and NSTX. This unique code was developed to model multiple solid and liquid phase materials with arbitrary geometry. The inclusion of complex-geometry, electrically conducting walls and nozzles are essential since electric current closure paths are typically through these solid structures. In HIMAG, a second-order variable density projection method is used to simulate incompressible Navier–Stokes equations and the level set method is used to capture free surfaces. HIMAG is developed on unstructured grids, and can be run in parallel across multiple processors, and is thus able to efficiently solve large complex problems. HIMAG has already been validated for steady and unsteady single-fluid flow with/without MHD effects. In this paper, we demonstrate some validation cases for interfacial flows using HIMAG.}
+}
 
-
+@Manual{COMSOLRef,
+  title = {{COMSOL Multiphysics® Reference Manual}},
+  version = {6.3},
+  publisher = {{COMSOL AB}},
+  address = {{Stockholm, Sweden}},
+  year = {2024},
+  url = {www.comsol.com}
+}
