### 2025-09-01 14aaa8f
Update paper.md
Included references
