+++ paper/paper.md
@@ -51,7 +51,7 @@ The demand for high-performance computational fluid dynamics and multiphysics so
 
 # State of the field
 
-Magnetohydrodynamic (MHD) simulations can be performed using both HIMAG and COMSOL Multiphysics, though each platform addresses different computational objectives. HIMAG is a specialized solver optimized for high-fidelity modeling of liquid-metal flows in complex geometries under strong magnetic fields, such as those encountered in fusion and industrial applications. Its design provides enhanced computational efficiency and accuracy for these niche MHD problems but limits flexibility for coupling with additional physical phenomena. In contrast, COMSOL Multiphysics is a general-purpose multiphysics environment that enables the simultaneous solution of fluid dynamics, electromagnetics, heat transfer, and structural mechanics. This versatility allows the simulation of a broad spectrum of MHD applications, including electromagnetic stirring, liquid-metal batteries, and reactor-relevant flows. Nevertheless, COMSOL’s generalist approach may present challenges in achieving numerical convergence, computational efficiency, and stable solver performance in extreme MHD regimes found in large Hartmann flows. Overall, HIMAG is particularly suited for targeted, high-fidelity fusion-scale MHD simulations, whereas COMSOL provides a flexible framework for investigating MHD phenomena within coupled multiphysics contexts. Additionally, demonstration of scalibility of HIMAG and COMSOL Multiphysics on GPU HPC platforms is yet to be demonstrated.
+Magnetohydrodynamic (MHD) simulations can be performed using both HIMAG [@HIMAG] and COMSOL Multiphysics [@COMSOLRef], though each platform addresses different computational objectives. HIMAG is a specialized solver optimized for high-fidelity modeling of liquid-metal flows in complex geometries under strong magnetic fields, such as those encountered in fusion and industrial applications. Its design provides enhanced computational efficiency and accuracy for these niche MHD problems but limits flexibility for coupling with additional physical phenomena. In contrast, COMSOL Multiphysics is a general-purpose multiphysics environment that enables the simultaneous solution of fluid dynamics, electromagnetics, heat transfer, and structural mechanics. This versatility allows the simulation of a broad spectrum of MHD applications, including electromagnetic stirring, liquid-metal batteries, and reactor-relevant flows. Nevertheless, COMSOL’s generalist approach may present challenges in achieving numerical convergence, computational efficiency, and stable solver performance in extreme MHD regimes found in large Hartmann flows. Overall, HIMAG is particularly suited for targeted, high-fidelity fusion-scale MHD simulations, whereas COMSOL provides a flexible framework for investigating MHD phenomena within coupled multiphysics contexts. Additionally, demonstration of scalibility of HIMAG and COMSOL Multiphysics on GPU HPC platforms is yet to be demonstrated.
 
 # Statement of need
 
@@ -121,6 +121,7 @@ The authors declare that they have no known competing financial interests or per
 This work was funded by the Laboratory Directed Research and Development (LDRD) program at Oak Ridge National Laboratory. This manuscript has been authored by UT-Battelle, LLC, under contract DE-AC05-00OR22725 with the US Department of Energy (DOE). The US government retains and the publisher, by accepting the article for publication, acknowledges that the US government retains a nonexclusive, paid-up, irrevocable, worldwide license to publish or reproduce the published form of this manuscript, or allow others to do so, for US government purposes. DOE will provide public access to these results of federally sponsored research in accordance with the [DOE Public Access Plan](http://energy.gov/downloads/doe-public-access-plan).
 
 # References
+
 
 
 
