### 2025-09-01 184e440
adding figures with results
