### 2025-09-02 51b2a09
Update paper.md
Update text
