+++ paper/paper.md
@@ -61,15 +61,13 @@ Two approaches are commonly adopted to predict MS flows exposed to a magnetic fi
 
 VERTEX-CFD is a new open-source package designed to address the modeling of MHD flows in complex geometries. It provides a robust multi-physics solver that can generate high-fidelity simulations by scaling on HPC platforms. Artificial intelligence and machine learning (AI&ML) capabilities are provided by TensorFlow Lite package [@tensorflow2015-whitepaper] that is coupled to VERTEX-CFD. The coupling allows to enhance physical models with ML models at runtime. These capabilities will provide the needed resources to extend the applicability of turbulence models to MHD flows, and perform modeling of complex flows under the influence of an external magnetic field. This in turn will be leveraged for the modeling of realistic complex geometries found in fusion blankets to capture the coupled multiphysics behavior that governs blanket performance and lifetime.
 
-The need for scalability on HPC platforms and flexibility of source code led the authors to investigate different open-source high-performace computing tools: MFEM [@mfem], Trilinos [@trilinos-website] and Nek5000[@nek5000]. MFEM is a lightweight, flexible library focused on high-order finite element discretizations, ideal for prototyping new PDE solvers and coupling with external linear algebra packages. Trilinos is a broad collection of modular packages that provides scalable solvers, discretizations, and multiphysics coupling capabilities, making it a comprehensive but complex toolkit for large scientific codes. Nek5000, in contrast, is a specialized, production-ready CFD solver based on the spectral element method, optimized for incompressible flows and turbulence simulations. Trilinos was selected as the base for the open-source VERTEX-CFD package due to its comprehensive suite of scalable solvers, advanced preconditioning techniques, robust multiphysics coupling capabilities, and the familiartity of the authors with the source code base. Its modular design and mature software infrastructure provide a flexible foundation for integrating diverse discretization strategies and targeting emerging high-performance architectures. By leveraging Trilinos, the package benefits from a well-established ecosystem that has been widely adopted in large-scale scientific applications, ensuring both reliability and extensibility for future developments.
-
 # Current capabilities and development workflow
 
-![Caption for example figure.\label{fig:example}](figures/VERTEX-CFD-one-slide.png)
+The need for scalability on HPC platforms and flexibility of source code led the authors to investigate different open-source high-performace computing tools: MFEM [@mfem], Trilinos [@trilinos-website] and Nek5000[@nek5000]. MFEM is a lightweight, flexible library focused on high-order finite element discretizations, ideal for prototyping new PDE solvers and coupling with external linear algebra packages. Trilinos is a broad collection of modular packages that provides scalable solvers, discretizations, and multiphysics coupling capabilities, making it a comprehensive but complex toolkit for large scientific codes. Nek5000, in contrast, is a specialized, production-ready CFD solver based on the spectral element method, optimized for incompressible flows and turbulence simulations. Trilinos was selected as the base for the open-source VERTEX-CFD package due to its comprehensive suite of scalable solvers, advanced preconditioning techniques, robust multiphysics coupling capabilities, and the familiartity of the authors with the source code base. Its modular design and mature software infrastructure provide a flexible foundation for integrating diverse discretization strategies and targeting emerging high-performance architectures. By leveraging Trilinos, the package benefits from a well-established ecosystem that has been widely adopted in large-scale scientific applications, ensuring both reliability and extensibility for future developments.
 
[…]
