### 2025-09-02 ORNL/VERTEX-CFD#9 @delcmo
@sthavishtha thanks for your review. Please find my responses below to your comments:

1. The list of authors is correct. The package has been developed on GitLab repo hosted on the ORNL network. It has been made open source for a few months now after receiving approval from DOE.
2. I added a figure that has some numerical results of benchmark problems that were used to V&V the implementation.
3. At the time we submitted the paper, RANS turbulence models were under development but are now available.
4. I updated the text with the following statement:
_These capabilities will provide the needed resources to extend the applicability of turbulence models to MHD flows, and perform modeling of flows under the influence of an external magnetic field. This in turn will be leveraged for the modeling of realistic complex geometries found in fusion blankets to capture the coupled multiphysics behavior that governs blanket performance and lifetime._
5. I added some more content:
_Artificial intelligence and machine learning (AI&ML) capabilities are provided by TensorFlow Lite package [@tensorflow2015-whitepaper] that is coupled to VERTEX-CFD. The coupling allows to enhance physical models with ML models at runtime._
6. I removed it and replaced it with new content. See response to 4.
7. I added a paragraph in **Current capabilities and development workflow** section
_The need for scalability on HPC platforms and flexibility of source code led the authors to investigate different open-source high-performace computing tools: MFEM [@mfem], Trilinos [@trilinos-website] and Nek5000[@nek5000]. MFEM is a lightweight, flexible library focused on high-order finite element discretizations, ideal for prototyping new PDE solvers and coupling with external linear algebra packages. Trilinos is a broad collection of modular packages that provides scalable solvers, discretizations, and multiphysics coupling capabilities, making it a comprehensive but complex toolkit for large scientific codes. Nek5000, in contrast, is a specialized, production-ready CFD solver based on the spectral element method, optimized for incompressible flows and turbulence simulations. Trilinos was selected as the base for the open-source VERTEX-CFD package due to its comprehensive suite of scalable solvers, advanced preconditioning techniques, robust multiphysics coupling capabilities, and the familiartity of the authors with the source code base. Its modular design and mature software infrastructure provide a flexible foundation for integrating diverse discretization strategies and targeting emerging high-performance architectures. By leveraging Trilinos, the package benefits from a well-established ecosystem that has been widely adopted in large-scale scientific applications, ensuring both reliability and extensibility for future developments._
9. I added a **State of the field**.
10. I fixed the first link, thanks. I could not find  https://doi.org/  on line 173. The other link seems to work fine for me.
