### 2025-09-02 aa9020f
Update paper.md
