### 2025-09-03 ORNL/VERTEX-CFD#9 @sthavishtha
where is the link to the updated software paper. I am looking at https://github.com/openjournals/joss-papers/blob/joss.08654/joss.08654/10.21105.joss.08654.pdf and it isn't updated
