### 2025-09-03 ORNL/VERTEX-CFD#9 @delcmo
@sthavishtha the updated software paper can be found here https://github.com/ORNL/VERTEX-CFD/actions/runs/17390903388 in the artifact section at the bottom.
