### 2025-09-03 ORNL/VERTEX-CFD#9 @sthavishtha
these references are still not fixed:
- doi for "Delchini, M.O., Kincaid, K. C.,,......" gives me an "invalid doi error" link (this is line 147 in the older version of the paper)
- doi for the article "Smolentsev, S., Badia, S.,....." has an extra "https://doi.org/" - this has to be fixed. This is line 173 in the older version of the paper.
