### 2025-09-04 c849a45
Update paper.bib
