+++ paper/paper.bib
@@ -137,7 +137,7 @@ volume = {100},
 pages = {65-72},
 year = {2015},
 issn = {0920-3796},
-doi = {https://doi.org/10.1016/j.fusengdes.2014.04.049},
+doi = {https://doi-org.ornl.idm.oclc.org/10.1016/j.fusengdes.2014.04.049},
 url = {https://www.sciencedirect.com/science/article/pii/S0920379614003263},
 author = {S. Smolentsev and S. Badia and R. Bhattacharyay and L. Bühler and L. Chen and Q. Huang and H.-G. Jin and D. Krasnov and D.-W. Lee and E. Mas {de les Valls} and C. Mistrangelo and R. Munipalli and M.-J. Ni and D. Pashkevich and A. Patel and G. Pulugundla and P. Satyamurthy and A. Snegirev and V. Sviridov and P. Swain and T. Zhou and O. Zikanov},
 keywords = {Blanket, Liquid metal magnetohydrodynamics, Computer code},
@@ -262,7 +262,7 @@ abstract = {On the pathway toward full simulations for a liquid metal (LM) blank
 title = "VERTEX-CFD: A MULTIPHYSICS SOLVER",
 author = "Delchini, {Marc Olivier} and Kincaid, {Kellis C.} and Doug Stefanski and Ryan Glasby and Franklin Curtis",
 year = "2024",
-doi = "10.13182/T130-44950",
+doi = "",
 language = "English",
 volume = "130",
 pages = "1191--1194",
@@ -296,3 +296,4 @@ abstract = {HIMAG, a 3-D incompressible MHD free surface code developed by HyPer
   year = {2024},
   url = {www.comsol.com}
 }
+
