### 2025-09-04 cfd2f73
Update paper.bib
