### 2025-09-04 f78e27d
Update paper.bib
