+++ paper/paper.bib
@@ -137,7 +137,7 @@ volume = {100},
 pages = {65-72},
 year = {2015},
 issn = {0920-3796},
-doi = {https://doi.org/10.1016/j.fusengdes.2014.04.049},
+doi = {10.1016/j.fusengdes.2014.04.049},
 url = {https://www.sciencedirect.com/science/article/pii/S0920379614003263},
 author = {S. Smolentsev and S. Badia and R. Bhattacharyay and L. Bühler and L. Chen and Q. Huang and H.-G. Jin and D. Krasnov and D.-W. Lee and E. Mas {de les Valls} and C. Mistrangelo and R. Munipalli and M.-J. Ni and D. Pashkevich and A. Patel and G. Pulugundla and P. Satyamurthy and A. Snegirev and V. Sviridov and P. Swain and T. Zhou and O. Zikanov},
 keywords = {Blanket, Liquid metal magnetohydrodynamics, Computer code},
@@ -296,5 +296,6 @@ abstract = {HIMAG, a 3-D incompressible MHD free surface code developed by HyPer
   year = {2024},
   url = {www.comsol.com}
 }
+
 
 
