### 2025-09-04 ORNL/VERTEX-CFD#9 @delcmo
@sthavishtha, I fixed the references:
- I removed the doi link for 'Delchini, M.O., Kincaid, K. C' as the link was invalid. I cannot find the correct one.
- I removed the extraneous `https`.
Thanks agains,
Marco
