### 2025-09-29 ORNL/VERTEX-CFD#10 @Phoenixfire1081
Review of article: VERTEX-CFD
Here are my comments on the [article](https://github.com/ORNL/VERTEX-CFD/actions/runs/17390903388),

- L62: High-fidelity simulation - why is LES included in this? As far as what I know, the quality of LES simulations depends on the underlying sub-grid scale models.
- L66-L67: Instead of "The main technique that reduces", it can be written as "One of the techniques that reduces.." since LES also reduces computational requirements.
- L126: Perhaps "for scalability" is better than "to scale".
- Figure 1 b, c, e - Legends and axis lables are too small to be readable.
- Since scalability appears to be one of the main features of the software, perhaps a plot showing it would be nice.

The paper is well-written and addresses requirements of the JOSS checklist quite well. I will report on my second set of comments focusing on the documentation and functionality part of the checklist.
