### 2025-09-29 #8654 @Phoenixfire1081
Here are my first set of [comments](https://github.com/ORNL/VERTEX-CFD/issues/10#issue-3463396555).
