### 2025-09-30 1c3d823
Update paper.md
