+++ paper/paper.md
@@ -57,7 +57,7 @@ Magnetohydrodynamic (MHD) simulations can be performed using both HIMAG [@HIMAG]
 
 Fusion energy development draws from a wide range of disciplines to describe design and to develop a functioning system. One challenging engineering task is to develop the fusion core component known as a "blanket." Because this component surrounds the burning plasma and must absorb almost all of the power from nuclear reactions, it must breed fuel, provide nuclear shielding, and provide energy deposition. Molten salt (MS) is a primary choice for cooling the blanket. A “salt blanket” in fusion energy is a layer of molten salt surrounding the fusion plasma. The molten salt acts as both a coolant and a material for neutron absorption, both of which are essential in fusion reactions. The salt blanket absorbs the high-energy neutrons produced by fusion, reducing the wear on reactor components, and converting some of the energy into heat for electricity generation. Molten salts have low electrical and thermal conductivity and experience lesser electromagnetic forces, but they are still turbulent. Heat transfer degradation in an MS flow caused by the reduction of turbulence by a magnetic field is a possible limitation of the MS blanket [@Smolentsev01042005].
 
-Two approaches are commonly adopted to predict MS flows exposed to a magnetic field: high-fidelity simulation (large-eddy simulation (LES) or direct numerical simulation (DNS)), and Reynolds-averaged Navier-Stokes (RANS) turbulence models. LESs can resolve turbulences at temporal and spatial scales at the expense of large HPC resources. Blanket design with LES is not possible because of current HPC limitations. Design optimization often requires multiple simulation runs to investigate performance under various conditions. The main technique that reduces the computational requirements of the analysis is the RANS turbulence model. This approach filters out the instantaneous velocity component, and the influence of the turbulence is modeled solely by the closure models. Turbulence modeling is a complex problem, and many turbulence models are available as described in the literature [@Chen_2022; @Menter1992ImprovedTK], albeit with many limitations [@10.1023/a:1022818327584]. Furthermore, these models are not readily applicable to the MHD flows and would require modifications [@Smolentsev2002] because MHD effects introduce additional terms in the turbulence balance equations.
+Two approaches are commonly adopted to predict MS flows exposed to a magnetic field: high-fidelity simulation (large-eddy simulation (LES) or direct numerical simulation (DNS)), and Reynolds-averaged Navier-Stokes (RANS) turbulence models. LESs can resolve turbulences at temporal and spatial scales at the expense of large HPC resources. Blanket design with LES is not possible because of current HPC limitations. Design optimization often requires multiple simulation runs to investigate performance under various conditions. One of the techniques that reduces the computational requirements of the analysis is the RANS turbulence model. This approach filters out the instantaneous velocity component, and the influence of the turbulence is modeled solely by the closure models. Turbulence modeling is a complex problem, and many turbulence models are available as described in the literature [@Chen_2022; @Menter1992ImprovedTK], albeit with many limitations [@10.1023/a:1022818327584]. Furthermore, these models are not readily applicable to the MHD flows and would require modifications [@Smolentsev2002] because MHD effects introduce additional terms in the turbulence balance equations.
 
[…]
