### 2025-09-30 ORNL/VERTEX-CFD#10 @delcmo
@Phoenixfire1081 thanks for providing comments. Below are my answers to your comments:

- **L62: High-fidelity simulation - why is LES included in this? As far as what I know, the quality of LES simulations depends on the underlying sub-grid scale models.**
LES techniques are commonly used to generate high-fidelity simulations of flow in complex geometries. In the context of MHD flows this is still valid under the assumption that the LES methods is adapted to account for the MHD effects.

- **L66-L67: Instead of "The main technique that reduces", it can be written as "One of the techniques that reduces.." since LES also reduces computational requirements.**
Thanks for the suggestion. We updated the text.

- **L126: Perhaps "for scalability" is better than "to scale"**.
Thanks for the suggestion. We updated the text.

- **Figure 1 b, c, e - Legends and axis lables are too small to be readable.**
I will re-generate the plots and get back to you. Thanks.

- **Since scalability appears to be one of the main features of the software, perhaps a plot showing it would be nice.**
Plot (c) in Figure 1 shows strong scaling of the VERTEX-CFD solver on Frontier.

The updated version can be found in here: https://github.com/ORNL/VERTEX-CFD/actions/runs/18129149340

Thanks,

Marco
