### 2025-09-30 ORNL/VERTEX-CFD#10 @delcmo
@Phoenixfire1081 I updated the figure and the latest version of the manuscript can be found here: https://github.com/ORNL/VERTEX-CFD/actions/runs/18138001609

Marco
