### 2025-10-01 ORNL/VERTEX-CFD#10 @Phoenixfire1081
Thanks for your efforts. I'll get back to you with my second set of comments soon.
