+++ src/incompressible_lsvof_solver/closure_models/VertexCFD_Closure_IncompressibleLSVOFErrorNorms_impl.hpp
@@ -1,0 +1,135 @@
+#ifndef VERTEXCFD_CLOSURE_INCOMPRESSIBLELSVOFERRORNORMS_IMPL_HPP
+#define VERTEXCFD_CLOSURE_INCOMPRESSIBLELSVOFERRORNORMS_IMPL_HPP
+
+#include "utils/VertexCFD_Utils_VectorField.hpp"
+
+#include <Panzer_GlobalIndexer.hpp>
+#include <Panzer_HierarchicParallelism.hpp>
+#include <Panzer_PureBasis.hpp>
+#include <Panzer_Workset_Utilities.hpp>
+
+#include <Teuchos_Array.hpp>
+
+#include <cmath>
+#include <string>
+
+namespace VertexCFD
+{
+namespace ClosureModel
+{
+//---------------------------------------------------------------------------//
+template<class EvalType, class Traits, int NumSpaceDim>
+IncompressibleLSVOFErrorNorms<EvalType, Traits, NumSpaceDim>::
+    IncompressibleLSVOFErrorNorms(const panzer::IntegrationRule& ir,
+                                  const Teuchos::ParameterList& closure_params)
+    : _L1_error_continuity("L1_Error_continuity", ir.dl_scalar)
+    , _L2_error_continuity("L2_Error_continuity", ir.dl_scalar)
+    , _volume("volume", ir.dl_scalar)
+    , _dof_name(closure_params.get<std::string>("DOF Name"))
+    , _exact_pressure("Exact_lagrange_pressure", ir.dl_scalar)
+    , _exact_dof("Exact_" + _dof_name, ir.dl_scalar)
+    , _pressure("lagrange_pressure", ir.dl_scalar)
+    , _dof(_dof_name, ir.dl_scalar)
+    , _L1_error_dof("L1_Error_" + _dof_name, ir.dl_scalar)
+    , _L2_error_dof("L2_Error_" + _dof_name, ir.dl_scalar)
+{
+    //  Check if NS equations are solved
+    _solve_ns = closure_params.isType<bool>("Compute NS Error Norms")
+                    ? closure_params.get<bool>("Compute NS Error Norms")
+                    : false;
+
+    // Add NS terms
+    if (_solve_ns)
+    {
+        // Dependent fields
+        this->addDependentField(_exact_pressure);
+        this->addDependentField(_pressure);
+
+        Utils::addDependentVectorField(
+            *this, ir.dl_scalar, _exact_velocity, "Exact_velocity_");
+        Utils::addDependentVectorField(
+            *this, ir.dl_scalar, _velocity, "velocity_");
+
+        // Evaluated fields
+        this->addEvaluatedField(_L1_error_continuity);
+        this->addEvaluatedField(_L2_error_continuity);
+
+        Utils::addEvaluatedVectorField(
+            *this, ir.dl_scalar, _L1_error_momentum, "L1_Error_momentum_");
+        Utils::addEvaluatedVectorField(
+            *this, ir.dl_scalar, _L2_error_momentum, "L2_Error_momentum_");
+    }
+
+    // LSVOF terms
+    this->addDependentField(_dof);
+    this->addDependentField(_exact_dof);
+    this->addEvaluatedField(_L1_error_dof);
+    this->addEvaluatedField(_L2_error_dof);
+
+    this->addEvaluatedField(_volume);
+
+    this->setName("Incompressible LSVOF Error Norms "
+                  + std::to_string(num_space_dim) + "D");
+}
+
+//---------------------------------------------------------------------------//
+template<class EvalType, class Traits, int NumSpaceDim>
+void IncompressibleLSVOFErrorNorms<EvalType, Traits, NumSpaceDim>::evaluateFields(
[…]
