+++ src/incompressible_lsvof_solver/closure_models/VertexCFD_IncompressibleLSVOFClosureModelFactory_Tangent3d.cpp
@@ -1,0 +1,5 @@
+#include "VertexCFD_IncompressibleLSVOFClosureModelFactory.hpp"
+#include "VertexCFD_IncompressibleLSVOFClosureModelFactory_impl.hpp"
+
+template class VertexCFD::ClosureModel::
+    IncompressibleLSVOFFactory<panzer::Traits::Tangent, 3>;
