+++ src/closure_models/VertexCFD_Closure_VariableTauSUPG.cpp
@@ -1,0 +1,7 @@
+#include "utils/VertexCFD_Utils_ExplicitTemplateInstantiation.hpp"
+
+#include "VertexCFD_Closure_VariableTauSUPG.hpp"
+#include "VertexCFD_Closure_VariableTauSUPG_impl.hpp"
+
+VERTEXCFD_INSTANTIATE_TEMPLATE_CLASS_EVAL_TRAITS_NUMSPACEDIM(
+    VertexCFD::ClosureModel::VariableTauSUPG)

[…]
