### 2025-12-05 ORNL/VERTEX-CFD#10 @Phoenixfire1081
Here are my second set of comments:

- The installation instructions are not very clear especially the Trilinos part - I wasn't even table to install the Trilinos-CPU. I tried installing the software on a Ubuntu 24.04 machine with NVIDIA RTX 6000 Ada Generation. It would be helpful to know what machines/OS the authors have tested the software on so that common issues during installation can be mentioned in the documentation itself. Due to this reason, I've been unable to test the functionality of the software. After the authors report the OS they've tested, I'll use a virtual machine with the same OS and test the functionality of the software.
- While the included example seems fine, additional examples in the documentation are needed so that the user has an easier time understanding the software.
- Did I miss it, or is there a documentation of the API?
