### 2025-12-07 ORNL/VERTEX-CFD#10 @delcmo
@Phoenixfire1081 thanks for the feedback.

- I will add some additional content to the installation instruction section.
- I will add one or two additional examples.
- What do you mean by API? An API to generate the XML input file?
