### 2026-01-20 #8654 @Phoenixfire1081
@piyueh Despite my attempts at installing the software, I've been so far unsuccessful. The problem isn't with the software under review but its dependency called Trilinos. I have spent few hours trying to debug but my attempts ended in vain. At this point, I would like to ask you how to proceed. From my checklist, only the functionality part remains.
