### 2026-01-20 #8654 @piyueh
Thanks for the feedback, @Phoenixfire1081. Could you please share some details about the machine and operating system you used to install and test the software (and any available log files)? This would help the author reproduce the issue you encountered.

@delcmo: Is the installation documentation intended only for some specific HPC clusters and systems? Does the software also support personal computers or machines with different management systems (for example, those don't use `module` for environment management)? If so, it might be helpful to provide more general installation instructions (e.g., for personal computers and common operating systems).
