### 2026-01-20 #8654 @delcmo
@Phoenixfire1081 it would be helpful if you can share the configuration script used to install Trilinos and any error messages. Did you try with the configuration scripts I shared with you a couple of weeks ago? Do you have by any chance access to Frontier (OLCF)?

@piyueh we have installed the software on personal computers and HPC machines. I can try to provide more general information but I would need feedback from the reviewers on the error messages and what the steps they followed.

Marco
