### 2026-01-22 #8654 @Phoenixfire1081
@piyueh @delcmo I apologize for the late reply, there were a couple of deadlines. I'll reply actively from now until the issue is resolved and the review is completed.

@delcmo No, I do not have access to Frontier.

Regarding my tests: I followed the instructions provided [here](https://ornl.github.io/VERTEX-CFD/docs/installation/).
Everything worked fine until spack install.

I originally used the configuration file shown in the installation webpage. I attempted to modify it based on the errors that were thrown - I didn't save the logs but from what I remember the issue was with a package called kokkos. Regardless, after @delcmo provided the new configuration file, I used that and most issues were resolved. However, trilinos still fails to install. I've attached the output of spack install here.

[spack_output.txt](https://github.com/user-attachments/files/24790966/spack_output.txt)

Please note that my .yaml script is similar to @delcmo. I simply removed the compilers part and let it automatically set the paths to gcc, g++, gfortran. Also, I set the concretizer to unify: true.

spack:
  specs:
    - gcc@11.2.0%gcc@8.5.0
    - cmake@3.27.9
    - ninja@1.10.1
    - boost@1.85.0
    - openmpi@4.1.1+legacylaunchers+pmi+internal-pmix
    - zlib@1.3.1
    - hdf5@1.14.3+cxx+fortran
    - netcdf-c@4.9.2+parallel-netcdf^openmpi@4.1.1+legacylaunchers+pmi+internal-pmix^hdf5@1.14.3 ^zlib@1.3.1
    - parmetis@4.0.3^openmpi@4.1.1+legacylaunchers+pmi+internal-pmix
    - cgns@4.4.0^openmpi@4.1.1+legacylaunchers+pmi+internal-pmix^hdf5@1.14.3
    - googletest@1.10.0

  concretizer:
    unify: true

I'm attempting to install the software on my workstation:
OS: Ubuntu 24.04
CPU: Intel Xeon w5 x 24
GPU: NVIDIA RTX 6000 Ada Generation
RAM: 128GB
We have a lot of disk storage, so that's not a problem.
