### 2026-01-22 #8654 @delcmo
@Phoenixfire1081 the below spack environment was used to install Trilinos on an Ubunto 22.4 machine:

```
# This is a Spack Environment file.
#
# It describes a set of packages to be installed, along with
# configuration settings.
spack:
  # add package specs to the `specs` list
  specs:
  - gcc@11.2.0%gcc@11.4.0
  - gcc@11.2.0
  - googletest%gcc@11.2.0
  - qt%gcc@11.2.0
  - cmake%gcc@11.2.0
  - ninja%gcc@11.2.0
  - openmpi@4.1.1%gcc@11.2.0
  - boost@1.74.0%gcc@11.2.0
  - netcdf-c@4.7.4%gcc@11.2.0
  - parmetis@4.0.3%gcc@11.2.0
  view: true
  concretizer:
    unify: true
  compilers:
  - compiler:
      spec: gcc@=12.3.0
      paths:
        cc: /usr/bin/gcc-12
        cxx: null
        f77: null
        fc: null
      flags: {}
      operating_system: ubuntu22.04
      target: x86_64
      modules: []
      environment: {}
      extra_rpaths: []
  - compiler:
      spec: gcc@=11.2.0
      paths:
        cc: /home/wll/build/spack/var/spack/environments/vertex_cfd_dev/.spack-env/view/bin/gcc
        cxx: /home/wll/build/spack/var/spack/environments/vertex_cfd_dev/.spack-env/view/bin/g++
        f77: /home/wll/build/spack/var/spack/environments/vertex_cfd_dev/.spack-env/view/bin/gfortran
        fc: /home/wll/build/spack/var/spack/environments/vertex_cfd_dev/.spack-env/view/bin/gfortran
      flags: {}
      operating_system: ubuntu22.04
      target: x86_64
      modules: []
      environment: {}
      extra_rpaths: []
```
You may have to edit some of the path to local compilers.

Trilinos is compiled separately after installing all dependencies with the spack file using the configuration script:
```
MPI_BIN_DIR=/home/wll/build/spack/var/spack/environments/vertex_cfd_dev/.spack-env/view/bin

cmake \
  -G Ninja \
  -DCMAKE_POLICY_VERSION_MINIMUM=3.5 \
  -DTPL_ENABLE_MPI=ON \
  -DBLAS_LIBRARY_DIRS=/usr/lib/x86_64-linux-gnu/blas \
  -DLAPACK_LIBRARY_DIRS=/usr/lib/x86_64-linux-gnu/lapack \
  -DCMAKE_C_COMPILER=$MPI_BIN_DIR/mpicc \
  -DCMAKE_CXX_COMPILER=$MPI_BIN_DIR/mpicxx \
  -DCMAKE_Fortran_COMPILER=$MPI_BIN_DIR/mpif90 \
  -DTrilinos_ENABLE_Fortran=OFF \
  -DBUILD_SHARED_LIBS=ON \
  -DTrilinos_ENABLE_ALL_PACKAGES:BOOL=OFF \
  -DTrilinos_ENABLE_Teuchos=ON \
  -DTrilinos_ENABLE_Tpetra=ON \
  -DTrilinos_ENABLE_Tempus=ON \
  -DTrilinos_ENABLE_Sacado=ON \
  -DTrilinos_ENABLE_Kokkos=ON \
  -DTrilinos_ENABLE_Panzer=ON \
  -DTrilinos_ENABLE_AztecOO=ON \
  -DTrilinos_ENABLE_SEACAS=ON \
  -DTrilinos_ENABLE_Ifpack=ON \
  -DTrilinos_ENABLE_ML=ON \
  -DTrilinos_ENABLE_Epetra=ON \
  -DTrilinos_ENABLE_EpetraExt=ON \
  -DTrilinos_ENABLE_ThyraTpetraAdapters=ON \
  -DTrilinos_ENABLE_ThyraEpetraAdapters=ON \
  -DTrilinos_ENABLE_ThyraEpetraExtAdapters=ON \
  -DTrilinos_ENABLE_Belos=ON \
  -DTrilinos_ENABLE_Stratimikos=ON \
  -DTrilinos_ENABLE_Amesos=ON \
  -DTrilinos_ENABLE_Amesos2=ON \
  -DPANZER_HAVE_EPETRA_STACK=ON \
  -DNOX_ENABLE_STRATIMIKOS_EPETRA_STACK=ON \
  -DTrilinos_ENABLE_TESTS=ON \
  -DTrilinos_SHOW_DEPRECATED_WARNINGS=OFF \
  -DCMAKE_BUILD_TYPE=Release \
  -DCMAKE_INSTALL_PREFIX=$TRILINOS_TARGET_DIR \
  -DCMAKE_EXPORT_COMPILE_COMMANDS:BOOL=ON \
  ../.
```

Let me know if it helps.

Marco
