### 2026-01-23 #8654 @Phoenixfire1081
@delcmo I used the new spack environment and it still doesn't work. There seems to be some issue with netCDF and exodus.

>> 3593    /home/ap269293/Documents/VERTEXCFD/trilinos/spack/var/spack/spack-stage/build-linux-ubuntu24.04-sapphirerapids-2026-01-23/spack-stage
             -trilinos-16.0.0-zfrwti6re7pbmpm7laet4zayu3xo4ze7/spack-src/packages/seacas/libraries/exodus/src/ex_put_prop.c:246:54: error: '_FillV
             alue' undeclared (first use in this function); did you mean 'NC_FillValue'?
     3594      246 |     if ((status = nc_put_att_longlong(exoid, propid, _FillValue, int_type, 1, vals)) != NC_NOERR) {
     3595          |                                                      ^~~~~~~~~~
     3596          |                                                      NC_FillValue

>> 3598    make[2]: *** [packages/seacas/libraries/exodus/CMakeFiles/exodus.dir/build.make:2336: packages/seacas/libraries/exodus/CMakeFiles/exo
             dus.dir/src/ex_put_prop.c.o] Error 1

The issue is probably that Trilinos has very specific dependencies like a specific version of netCDF needs to be installed or something similar.

However, I did manage to install it after another hour of debugging. The process was something like this:

- If there is an error in the installation, the spack.lock has to be removed and it is located at ../spack/var/spack/environments/vertex/
- There seems to be a copy of the spack.yaml file in ../spack/var/spack/environments/vertex/ and any changes after erroring needs to be made here directly.
- I had to modify the .yaml script. Here's what I used:

```
spack:
  specs:
  - gcc@13.3.0
  - googletest%gcc@13.3.0
  - qt%gcc@13.3.0
  - cmake%gcc@13.3.0
  - ninja%gcc@13.3.0
  - openmpi@4.1.1%gcc@13.3.0
  - boost@1.74.0%gcc@13.3.0
  - netcdf-c@4.7.4%gcc@13.3.0
  - parmetis@4.0.3%gcc@13.3.0
  - trilinos@16.1.0:%gcc@13.3.0
  view: true
  concretizer:
    unify: true
  compilers:
  - compiler:
      spec: gcc@=13.3.0
      paths:
        cc: /usr/bin/gcc
        cxx: /usr/bin/g++
        f77: /usr/bin/gfortran
        fc: /usr/bin/gfortran
      flags: {}
      operating_system: ubuntu24.04
      target: x86_64
      modules: []
      environment: {}
      extra_rpaths: []
  packages:
    gcc:
      externals:
      - spec: gcc@13.3.0 languages:='c,c++,fortran'
        prefix: /usr
        extra_attributes:
          compilers:
            c: /usr/bin/gcc
            cxx: /usr/bin/g++
            fortran: /usr/bin/gfortran
          flags: {}
          environment: {}
          extra_rpaths: []
```

- To get this spack config,

```gcc -v```

this tells the gcc version,

```
which gcc
which g++
which gfortran
```

for the paths. `- trilinos@16.1.0:%gcc@13.3.0` had to be specifically added to the spack, otherwise it defaults to trilinos@16.0.0 which has conflicts with netCDF.

At this point, I must say this process is not user friendly. I would like to see if the other reviewers experience such issues during the installation process.

I will continue with the VERTEXCFD installation and I will post here again if I encounter some other issue.
