### 2026-01-27 #8654 @delcmo
Our approach has been to provide a configuration script for Trilinos but I start to wonder if this is a good idea. Trilinos website has very limited installation instructions: https://trilinos.github.io/getting_started.html

Let me know if you encounter additional errors. Most of our users rely on HPC cluster and only a few developers have compiled it locally.

Marco
