### 2026-02-04 #8654 @Phoenixfire1081
@delcmo VERTEX-CFD installation requires [examples](https://github.com/ORNL/VERTEX-CFD/blob/master/examples) and [regression_test](https://github.com/ORNL/VERTEX-CFD/blob/master/regression_test) both of which shows file not found. Are these not made public?
