### 2026-02-04 #8654 @delcmo
See https://github.com/openjournals/joss-reviews/issues/8654#issuecomment-3128788644.
