### 2026-02-16 #8654 @piyueh
Hi @Phoenixfire1081 @khemgautam27 @sthavishtha,

I’m following up on the current status of the review. I understand there were some earlier issues related to software installation. Please let me know if those have not yet been resolved.

If you find that the documentation does not sufficiently guide you through the installation process, I would appreciate any specific suggestions on how the authors could improve it.

@Phoenixfire1081, regarding the examples and regression_test files, if the documentation is unclear about where to locate or download them, please feel free to suggest revisions that would make this clearer.

Thank you.
