### 2026-02-16 #8654 @Phoenixfire1081
Hi @piyueh,

The issues are mostly resolved. However, VERTEX-CFD still fails to install for me. The issue is somehow linked to Trilinos. I'll spend sometime this week and finish up the installation at least.

Regarding the documentation, I have many suggestions for the authors, however, I'll upload the final set of comments once the installation is successful and I can test some basic aspects of the software.
