### 2026-02-16 #8654 @piyueh
Thanks for the update, @Phoenixfire1081.

@delcmo @Phoenixfire1081 @khemgautam27 @sthavishtha Since installation has become a blocking issue, I recommend that @delcmo provide either a containerized solution (Singularity/Apptainer, Docker, etc.) or a self-contained packaging option (e.g., AppImage) for reviewers and future users. A container or self-contained package with all required dependencies and VERTEX-CFD preinstalled would likely be the most practical approach. This would allow reviewers and users to run and test the software without having to troubleshoot installation problems.

Bundling the dependencies would also let reviewers recompile or rebuild VERTEX-CFD (if they choose) in a controlled environment, without needing to resolve dependency conflicts.

This suggestion is intended to help move the submission forward. If reviewers are encountering installation barriers, it is likely that future users and JOSS readers will face similar difficulties. Resolving this issue is necessary in order to proceed with the submission.

Please let me know your thoughts.
