### 2026-02-16 #8654 @delcmo
What is the error message obtained when compiling Vertex-CFD against Trilinos?

Marco
