### 2026-03-23 #8654 @Phoenixfire1081
Hi @piyueh, thanks for the message. As I have stated last month, I still couldn't get VERTEX-CFD to install for me and as a result I've been unable to complete the review. I'm not sure going through the issues one by one and fixing them is the best approach. As the other two reviewers haven't responded in months, it is difficult to tell whether this is an issue I'm facing or it is a general one. At this point, I suggest one of the following paths:
- I suggest the authors heed your [advice](https://github.com/openjournals/joss-reviews/issues/8654#issuecomment-3909559999) and provide a containerized solution with all dependencies which would remove all installation issues.
- Find other reviewers who have access to the FRONTIER system which apparently has the software to test it.

@piyueh, I leave it to you to decide.
