### 2026-03-30 #8654 @sthavishtha
@piyueh  during the installation, i faced a similar issue as @Phoenixfire1081, even though i was installing on a HPC cluster. I had to use trilinos@16.2.0 as the default trilinos@16.0.0 failed to install successfully. My next step is to install Vertex-CFD. I will resume the installation in a few days and send another update by the end of this week.
