### 2026-04-13 #8654 @sthavishtha
the installation of VERTEX CFD package is throwing errors on my end. Looks like some files are missing in the source code.
my detailed review comments regarding installation are here: https://github.com/ORNL/VERTEX-CFD/issues/13
