### 2026-04-13 #8654 @piyueh
@delcmo As a reminder, multiple reviewers have encountered significant installation difficulties. The review cannot proceed until the installation issue is resolved.

If resolving the installation issue is not easy, I would strongly suggest providing an Apptainer/Singularity container solution. I believe Apptainer/Singularity is officially supported by most of HPC centers.
